theme switcher and context to change theme
